backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
mysql-connector-python
werkzeug
python-dotenv
cryptography
PyJWT
